dom/tests/mochitest/geolocation/test_optional_api_params.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=452566
-->
<head>
  <title>Test for Bug 452566</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="geolocation_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=452566">Mozilla Bug 452566</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

// ensure we are using the right testing provider
resume_geolocationProvider(function() {
  force_prompt(true, test1);
});

/** Test for Bug 452566 **/
function test1() {
  ok(navigator.geolocation, "Should have geolocation");

  var exception = null;
  try {
    navigator.geolocation.getCurrentPosition();
  } catch (ex) {
    if (!(ex instanceof TypeError)) {
      throw ex;
    }
    exception = ex;
  }
  ok(exception, "Should have got an exception");

  exception = null;
  try {
    navigator.geolocation.getCurrentPosition(function() {});
  } catch (ex) {
    if (!(ex instanceof TypeError)) {
      throw ex;
    }
    exception = ex;
  }
  ok(!exception, exception);

  exception = null;
  try {
    navigator.geolocation.getCurrentPosition(function() {}, function() {});
  } catch (ex) {
    if (!(ex instanceof TypeError)) {
      throw ex;
    }
    exception = ex;
  }
  ok(!exception, exception);

  exception = null;
  try {
    navigator.geolocation.getCurrentPosition(function() {}, function() {}, {});
  } catch (ex) {
    if (!(ex instanceof TypeError)) {
      throw ex;
    }
    exception = ex;
  }
  ok(!exception, exception);

  exception = null;
  try {
    navigator.geolocation.watchPosition();
  } catch (ex) {
    if (!(ex instanceof TypeError)) {
      throw ex;
    }
    exception = ex;
  }
  ok(exception, "Should have got an exception");

  exception = null;
  try {
    navigator.geolocation.watchPosition(function() {});
  } catch (ex) {
    if (!(ex instanceof TypeError)) {
      throw ex;
    }
    exception = ex;
  }
  ok(!exception, exception);

  exception = null;
  try {
    navigator.geolocation.watchPosition(function() {}, function() {});
  } catch (ex) {
    if (!(ex instanceof TypeError)) {
      throw ex;
    }
    exception = ex;
  }
  ok(!exception, exception);

  exception = null;
  try {
    navigator.geolocation.watchPosition(function() {}, function() {}, {});
  } catch (ex) {
    if (!(ex instanceof TypeError)) {
      throw ex;
    }
    exception = ex;
  }
  ok(!exception, exception);

  SimpleTest.finish();
}

</script>
</pre>
</body>
</html>
